requirements.txt
django-rql>=3.14
djangorestframework
flatten_json